Moved upower.*pp out of the upower directories
